Move blocks table row and header to dedicated files
