+++ src/components/screens/monitor/blocks/tableHeader.js
@@ -1,0 +1,29 @@
+import grid from 'flexboxgrid/dist/flexboxgrid.css';
+
+export default changeSort => ([
+  {
+    title: 'Height',
+    classList: `${grid['col-md-1']}`,
+    sort: () => changeSort('height'),
+  },
+  {
+    title: 'Date',
+    classList: `${grid['col-md-2']}`,
+  },
+  {
+    title: 'Transactions',
+    classList: `${grid['col-md-3']}`,
+  },
+  {
+    title: 'Generated by',
+    classList: `${grid['col-md-2']}`,
+  },
+  {
+    title: 'Amount',
+    classList: `${grid['col-md-2']}`,
+  },
+  {
+    title: 'Forged',
+    classList: `${grid['col-md-2']}`,
+  },
+]);
